back-end code/requirements.txt
Flask==2.0.1
Flask_Cors==3.0.10
Flask_RESTful==0.3.9
firebase_admin==5.2.0
python-dotenv==0.19.2
gunicorn==20.1.0